Implemented a simpler "idiom" for printout from a C++ API.
1. Fortran calls a C++ print_output function;
2. C++ sets up what to print out and stores it in a const char * variable, calculating the array length;
3. C++ calls a Fortran subroutine print_from_fortran that takes care of the printing.

+++ src/fortran_sources/fortran_module.F90
@@ -73,7 +73,7 @@ module fortran_module
 
         subroutine print_from_fortran(string, length)
                
-                integer :: length
+                integer(c_size_t) :: length
                 character(length) :: string
 
                 print *, "length = ", length
